refactor(core-filters): let the mongo cache backend own its collection
The filter_productId_cache collection and its indexes were provisioned by
FiltersCollection, the factory shared by every filter consumer, even though the
collection is the private storage of one pluggable cache backend. A shop running
a custom Redis/HTTP backend through the filtersSettings seam still got the Mongo
collection and indexes created underneath it.

makeMongoDBCache now creates and indexes the collection itself, and the
FilterProductIdCacheRecord type moves with it (re-exported from the package index
so the public surface is unchanged). FiltersCollection returns only the canonical
Filters and FilterTexts.

configureSettings instantiates the Mongo default lazily - only when a member is
not overridden - so a backend that overrides all three never provisions the Mongo
collection at all. Verified: a full custom backend leaves only `filters` and
`filter_texts`, no `filter_productId_cache`.

Behaviour-safe for the default backend: the collection is still created at module
init (configureSettings runs makeMongoDBCache before the rest of the module), just
from the backend that owns it rather than the shared factory.
